Little bit of Shelter restructuring.
