+++ frontend/src/shelter/Shelter.js
@@ -1,24 +1,55 @@
-import React from "react";
+import React, { useEffect, useState } from 'react';
+import axios from 'axios';
 import { Link } from 'raviger';
 import { ListGroup } from 'react-bootstrap'
-import { ShelterForm } from "./ShelterForms";
-import { ShelterTable } from "./ShelterTables";
+import { ShelterForm } from './ShelterForms';
+import Header from '../components/Header';
 
-const header_style = {
-  textAlign: "center",
-};
+function Shelter() {
 
-const Shelter = () => (
-  <ListGroup className="flex-fill p-5 h-50">
-    <Link href="/shelter/new">
-      <ListGroup.Item action>CREATE NEW SHELTER</ListGroup.Item>
-    </Link>
-    <Link href="/shelter/list">
-      <ListGroup.Item action>SEE ALL SHELTERS</ListGroup.Item>
-    </Link>
-  </ListGroup>
-)
+  const [data, setData] = useState({shelters: [],  isFetching: false});
 
+  // Hook for initializing data.
+  useEffect(() => {
+    let source = axios.CancelToken.source();
+    const fetchShelters = async () => {
+      setData({shelters: [], isFetching: true});
+      // Fetch Shelter data.
+      await axios.get('/shelter/api/shelter', {
+        cancelToken: source.token,
+      })
+      .then(response => {
+        setData({shelters: response.data, isFetching: false});
+      })
+      .catch(error => {
+        console.log(error.response);
+        setData({shelters: [], isFetching: false});
+      });
+    };
+    fetchShelters();
+    // Cleanup.
+    return () => {
+      source.cancel();
+    };
+  }, []);
+
+  return (
+    <>
+    <Header>Shelter</Header>
+    <hr/>
+    <ListGroup className="flex-fill p-5 h-50">
+      <Link href="/shelter/new">
+        <ListGroup.Item action>CREATE NEW SHELTER</ListGroup.Item>
+      </Link>
+      {data.shelters.map(shelter => (
+        <Link href={"/shelter/" + shelter.id}>
+          <ListGroup.Item action>{shelter.name} - {shelter.animal_count} Animals</ListGroup.Item>
+        </Link>
+      ))}
+    </ListGroup>
+    </>
+  )
+}
 export const NewShelter = () => (
   <div>
     <ShelterForm />
@@ -28,12 +59,6 @@ export const NewShelter = () => (
 export const UpdateShelter = ({id}) => (
   <div>
     <ShelterForm id={id}/>
-  </div>
-)
-
-export const ShelterList = () => (
-  <div>
-    <ShelterTable />
   </div>
 )
